satt: allow skip known benchmarks
check if we've got the result in database and if so, just skip it.
Use --skip-known-benchmarks=ID for that. The ID is the ID of the tool.
It is there because otherwise we'd need to wait for tool version from
benchmarks to identify the tool id. We may fix it later...